requirements.txt
streamlit
pymupdf
pdfplumber
python-docx
requests
pytesseract
Pillow
